Basic functionality
First commit.

Adds a webview element that javascript can be submitted to. A simple
console to display errors or console output on that webview, and a help
text at the very top describing how to work the program. All javascript
is sent through Beam's chat API on a channel chosen at app launch.